Cast unknown line special warning to (unsigned short)
So "Unknown special: 65535 ..." instead of "-1" will properly appear on vanilla E2M7.

+++ src/doom/p_setup.c
@@ -593,7 +593,7 @@ static void P_LoadLineDefs (int lump)
 	    char  badline[11];
 	    char *string;
         
-	    sprintf(badspec, "%d", ld->special);
+	    sprintf(badspec, "%d", (unsigned short) ld->special);
 	    sprintf(badline, "%d", i);
 	    string = M_StringJoin("Unknown special ", badspec, " at line ", badline, NULL);
 	    CRL_printf(string, false);
