+++ src/heretic/p_setup.c
@@ -352,7 +352,7 @@ static void P_LoadLineDefs(int lump)
             char  badline[11];
             char *string;
 
-            sprintf(badspec, "%d", ld->special);
+            sprintf(badspec, "%d", (unsigned short) ld->special);
             sprintf(badline, "%d", i);
             string = M_StringJoin("UNKNOWN SPECIAL ", badspec, " AT LINE ", badline, NULL);
             CRL_printf(string, false);
